+++ src/test/java/ua/com/foxminded/university/mapper/impl/FormOfEducationResponseMapperImplTest.java
@@ -4,14 +4,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import ua.com.foxminded.university.TestsContextConfiguration;
-import ua.com.foxminded.university.dto.FormOfEducationRequest;
 import ua.com.foxminded.university.dto.FormOfEducationResponse;
 import ua.com.foxminded.university.entity.FormOfEducation;
-import ua.com.foxminded.university.mapper.interfaces.FormOfEducationRequestMapper;
 import ua.com.foxminded.university.mapper.interfaces.FormOfEducationResponseMapper;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class FormOfEducationResponseMapperImplTest {
 
@@ -23,9 +20,9 @@ class FormOfEducationResponseMapperImplTest {
     {
         context = new AnnotationConfigApplicationContext(TestsContextConfiguration.class);
         formOfEducationResponseMapper = context.getBean(FormOfEducationResponseMapperImpl.class);
-        formOfEducation = FormOfEducation.builder().withId((long)1).withName("FormOfEducation 1").build();
+        formOfEducation = FormOfEducation.builder().withId(1L).withName("FormOfEducation 1").build();
         formOfEducationResponse = new FormOfEducationResponse();
-        formOfEducationResponse.setId((long)1);
+        formOfEducationResponse.setId(1L);
         formOfEducationResponse.setName("FormOfEducation 1");
     }
